+++ db/StoredProcedures/PollingStates_GetByKey.sql
@@ -1,0 +1,11 @@
+CREATE PROCEDURE [dbo].[PollingStates_GetByKey]
+  @key NVARCHAR(256)
+AS
+BEGIN
+  SET NOCOUNT ON;
+  SELECT
+    [Key],
+    [ETag]
+  FROM [dbo].[PollingStates]
+  WHERE [Key] = @key;
+END;